change lib to telebot

+++ app/internal/repository/user-repository.go
@@ -22,7 +22,7 @@ func NewRepository(db *sql.DB) UserRepository {
 	}
 }
 
-func (ur *UserRepository) UserExist(telegram_id int) bool {
+func (ur *UserRepository) UserExist(telegram_id int64) bool {
 	var exist bool
 	err := ur.db.QueryRow("SELECT EXISTS(SELECT 1 FROM users WHERE telegram_id = ?)", telegram_id).Scan(&exist)
 	if err != nil {
@@ -37,14 +37,14 @@ func (ur *UserRepository) UserExist(telegram_id int) bool {
 	return false
 }
 
-func (ur *UserRepository) CreateUser(telegram_id int, name string, message_count int) {
+func (ur *UserRepository) CreateUser(telegram_id int64, name string, message_count int) {
 	_, err := ur.db.Exec("INSERT INTO users (name, telegram_id, message_count) VALUES (?, ?, ?)", name, telegram_id, message_count)
 	if err != nil {
 		log.Print(err)
 	}
 }
 
-func (ur *UserRepository) AddUserMessageCount(telegram_id int) {
+func (ur *UserRepository) AddUserMessageCount(telegram_id int64) {
 	_, err := ur.db.Exec("UPDATE users SET message_count = message_count + 1 WHERE telegram_id = ?", telegram_id)
 	if err != nil {
 		log.Print(err)
